requirements.txt
python-dotenv
requests
pylint
psycopg2-binary
pandas